feat(trainer): consolidate routing + delete dead stub + resend on prereq unlock (#55)
Per the deep-dive in #55, the trainer NPC interaction was substantially
implemented but split across two competing code paths: the canonical
`template_trainer_lists` route (via `cell_methods/player/trainer_interaction.rs`)
and a dead `NpcInteractionType::Trainer { archetype_id }` arm in
`cell/interactions/dispatch.rs` that called a stub fabricating the ability list
with no per-known-ability / level / prereq filtering.

Changes:

1. Consolidate routing into a single canonical handler
   `cell::interactions::trainer::try_open_trainer`. The duplicate logic at
   `cell_methods/player/trainer_interaction.rs` is removed; the callsite in
   `cell_methods/player/interaction.rs` routes through the public path.
2. Convert the dead `NpcInteractionType::Trainer` arm into a warning-only
   stub — the variant isn't assigned anywhere in current content, but if
   future content edits set it, the canonical handler still keys off the
   template lookup. The prior stub `send_trainer_open` is deleted.
3. Delete the 42-line dead stub in `crates/game/src/interactions/trainer.rs`
   (`TrainerSession`, `TrainableAbility`, `TrainerResult` with `todo!()`).
   These types had no callers outside the module itself.
4. Add Python-parity resend-on-prereq-unlock behavior in
   `base_messages::AbilityGranted`: after `send_known_abilities_update`,
   re-send `onTrainerOpen` if the player has a trainer pinned via
   `last_interaction_target`. The recomputed `trainable` flags reflect the
   newly-known ability (a dependent ability becomes trainable once its
   prereq is learned). Mirrors `AbilityTrainer.onTrainAbility:128`.

Tests added:

- `trainer_interact_emits_exactly_one_on_trainer_open` (regression for
  the routing split — would have caught the double-emit if a future edit
  re-introduces it).
- `ability_granted_resends_trainer_open_with_unlocked_prereq` (positive:
  granting ability A (prereq of B) re-sends `onTrainerOpen` with B now
  trainable=1).
- `ability_granted_with_no_trainer_pinned_does_not_resend` (negative:
  chain-granted abilities don't spuriously open the trainer window).
- The existing `already_known_ability_shows_untrainable` regression is
  preserved from the consolidated module.

Wire format verified byte-exact per the deep-dive against
`entities/defs/SGWPlayer.def:1194-1198`, `alias.xml:417-422`, and the V5
RE findings at `docs/protocol/client-method-dispatch-table.md:248`.

Closes #55
