+++ crates/services/src/cell/cell_methods/player/mod.rs
@@ -4,9 +4,11 @@ pub mod crafting;
 mod dispatch;
 pub mod interaction;
 pub mod social;
-mod trainer_interaction;
 pub mod vendor;
 pub mod world;
+// `trainer_interaction` consolidated into
+// `cell::interactions::trainer::try_open_trainer` in #55 — single canonical
+// source-of-truth path. Callsite in `interaction.rs` routes through that.
 
 pub use constants::*;
 pub use dispatch::dispatch;
