+++ crates/services/src/cell/interactions/mod.rs
@@ -15,3 +15,4 @@ mod vendor;
 pub use dialog::send_dialog_display;
 pub use dispatch::{handle_initial_response, handle_interact};
 pub use loot::handle_loot_item;
+pub(crate) use trainer::try_open_trainer;
