+++ crates/services/src/cell/service/base_messages/mod.rs
@@ -379,6 +379,56 @@ pub(super) async fn handle_base_message(
                 "AbilityGranted: cell mirrored + hotbar refresh"
             );
             player_init::send_known_abilities_update(entity_id, tx, space_mgr).await;
+
+            // Python parity (`AbilityTrainer.onTrainAbility:128`): if the
+            // newly-learned ability is a prerequisite for another offered
+            // ability, OR the player just ran out of training points, the
+            // trainer list should refresh so the client's UI updates the
+            // greyed-out state. Without this, the player sees a stale list
+            // with the dependent ability still greyed out until they close
+            // and re-open the trainer.
+            //
+            // We delegate the "is this newly-unlocked a prereq for B?"
+            // decision to `try_open_trainer` itself — it recomputes every
+            // `trainable` flag from current state (known set, level,
+            // prereqs). Calling it whenever the player has a trainer pinned
+            // is idempotent and matches Python's "always-resend on grant"
+            // shape. `try_open_trainer` short-circuits to `false` when the
+            // pinned target isn't a trainer template, so the only NPCs
+            // that trigger a resend here are real trainers.
+            //
+            // `last_interaction_target` is set by `handle_interact` and
+            // not cleared on trainer close. Trade-off: if a player opens
+            // a trainer, closes it, then earns an ability some other way
+            // (chain `Action::GrantAbility` from a quest turn-in), we'd
+            // emit a spurious `onTrainerOpen`. The client tolerates an
+            // unsolicited `onTrainerOpen` when the trainer window isn't
+            // visible (UEvent_UI_TrainerOpen handler just shows the
+            // panel), so this is harmless. See issue #55 deep dive Item B.
+            let trainer_entity_id = space_mgr
+                .get_entity(entity_id)
+                .and_then(|p| p.last_interaction_target);
+            if let Some(target) = trainer_entity_id {
+                let is_trainer = space_mgr
+                    .get_entity(target)
+                    .and_then(|t| t.template_id)
+                    .is_some_and(|tid| space_mgr.template_trainer_lists.contains_key(&tid));
+                if is_trainer {
+                    tracing::debug!(
+                        target: "abilities",
+                        event = "trainer_resend",
+                        entity_id,
+                        ability_id,
+                        trainer_entity_id = target,
+                        training_points_remaining,
+                        "AbilityGranted: re-sending onTrainerOpen to refresh trainable flags"
+                    );
+                    let _ = crate::cell::interactions::try_open_trainer(
+                        entity_id, target, tx, space_mgr,
+                    )
+                    .await;
+                }
+            }
         }
 
         BaseToCellMsg::ItemUsed {
