+++ crates/services/src/cell/service/base_messages/tests/mod.rs
@@ -2,6 +2,7 @@ use super::*;
 use crate::cell::space_manager::SpaceManager;
 use cimmeria_entity::cell_entity::BandolierItem;
 
+mod ability_granted_trainer_resend;
 mod bandolier_sync;
 mod bandolier_update;
 mod general;
